requirements.txt
groq
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-generativeai
sentence-transformers
pandas
pinecone
tqdm
requests